Implements basic moving mean for lineplot.
Summary:

- Adds moving mean
- Mean window size hardcoded as 10
- Moved the line plotting code into its own function
    * This helps prepare for the user controlling
      the smoothing manually.

This commit in the moving-mean branch implements a basic
moving mean. Right now the mean window size is hardcoded
as size 10. This doesn't work well for small arrays
(e.g., 200 and probably more).

Next step is to give it some smarts or add a heuristic
for when to apply a moving mean and when not to.

Next next step is to give the user the option to enable
or disable the moving mean smoothing whenever they want.
And to specify the size of the window.