Smooth line plot tails and add weak smarts.
This continues work on the moving mean modification of the line
plot.

Now the tails of the line at either end are also smoothed,
and I added a not very smart heuristic for picking the moving
mean window size.

This is no alternative to the user being able to choose their
own.

On branch moving-mean
Your branch is up-to-date with 'origin/moving-mean'.

Changes to be committed:
      modified:   Support/HTML/byTime/byTime.js