This adds user override to moving mean win size.
This completes the smoothing/moving mean feature.

The user now has a text box that displays the size of the current
moving mean window. The user can choose a new value and click
"Redraw Line" to redraw the line plot with the specified moving
mean window.

I also added basic logic to check the value the user enters.

Protected against:

- nothing
- less than 0 numbers
- letters
- other alphanumeric combinations that do not resolve to an int

On branch moving-mean
Your branch is up-to-date with 'origin/moving-mean'.

Changes to be committed:
      modified:   byTime.js
      modified:   index.html